+++ R/ds.R
@@ -20,9 +20,10 @@
 #'  \item{DSSTDY}{Study Day of Start of Disposition Event}
 #'  \item{DSSCAT}{Subcategory for Disposition Event}
 #' }
-#' @seealso \code{\link{ds}} \code{\link{mh}}
+#' @seealso \code{\link{ds}} \code{\link{dv}} \code{\link{ie}} \code{\link{mh}}
 #' @keywords datasets ds
 #' @name ds
 #' @examples
-#' head(data("ds"))
+#'  head(data("ds"))
 "ds"
+