+++ R/dv.R
@@ -1,0 +1,32 @@
+#' @title dv
+#'
+#' @description dv modified from pharmaversesdtm
+#' @source data from pharmaversesdtm.
+#'
+#' @format A data frame with 75 rows and 17 variables:
+#' \describe{
+#'  \item{STUDYID}{Study Identifier}
+#'  \item{USUBJID}{Unique Subject Identifier}
+#'  \item{TRTSDTM}{Datetime of First Exposure to Treatment}
+#'  \item{TRTEDTM}{Datetime of Last Exposure to Treatment}
+#'  \item{DOMAIN}{Domain Abbreviation}
+#'  \item{DVSEQ}{Sequence Number}
+#'  \item{DVTERM}{Protocol Deviation Term}
+#'  \item{DVDECOD}{DVDECOD}
+#'  \item{DVCAT}{DVCAT}
+#'  \item{DVSCAT}{Subcategory for Protocol Deviation}
+#'  \item{DVREAS}{Reason for Deviation}
+#'  \item{DVEPRELI}{Epi/Pandemic Related Indicator}
+#'  \item{ASTDT}{Analysis Start Date}
+#'  \item{ASTDY}{Analysis Start Relative Day}
+#'  \item{AEPRELFL}{Analysis Epi/Pandemic Related Flag}
+#'  \item{TRTSDT}{Date of First Exposure to Treatment}
+#'  \item{DVSTDTC}{Date of First Exposure to Treatment}
+#' }
+#' @seealso \code{\link{ds}} \code{\link{dv}} \code{\link{ie}} \code{\link{mh}}
+#' @keywords datasets dv
+#' @name dv
+#' @examples
+#'  head(data("dv"))
+"dv"
+