+++ NEWS.md
@@ -1,2 +1,7 @@
-# pharmaversesdtmjnj 0.0.3
+# pharmaversesdtmjnj 0.0.2
+
+### Added
+* Added SDTM datasets that comply with J&J standards:
+  * DV (Deviation)
+  * IE (Inclusion/Exclusion)